requirements.txt
pandas==2.2.0
python-dotenv==1.0.1
SQLAlchemy==2.0.27

langchain~=0.1.7